+++ addons/website_sale/static/src/js/website_sale_tour_shop_custom_attribute_value.js
@@ -3,8 +3,6 @@ odoo.define("website_sale.tour_shop_custom_attribute_value", function (require) 
 
     var tour = require("web_tour.tour");
 
-    var optionVariantImage;
-
     tour.register("shop_custom_attribute_value", {
         url: "/shop?search=Customizable Desk",
         test: true,
@@ -19,59 +17,8 @@ odoo.define("website_sale.tour_shop_custom_attribute_value", function (require) 
         trigger: 'input.variant_custom_value',
         run: 'text Wood',
     }, {
+        id: 'add_cart_step',
         trigger: 'a:contains(Add to Cart)',
-        run: 'click',
-    }, {
-        trigger: 'div:contains(Custom: Wood)',
-        extra_trigger: '.js_product.in_cart.main_product',
-        run: function (){} // checks that Yep, it's wood!
-    }, {
-        trigger: 'button.js_add_cart_json:has(i.fa-plus)',
-        run: 'click',
-    }, {
-        trigger: 'div.oe_striked_price span:contains(750)',
-        run: function (){}, // check
-    }, {
-        trigger: 'span.oe_price span:contains(600)',
-        run: function (){}, // check
-    }, {
-        trigger: '.oe_optional_products_modal .js_product:eq(1) div:contains("Conference Chair (Steel)")',
-        run: function () {
-            optionVariantImage = $('.oe_optional_products_modal .js_product:eq(1) img.variant_image').attr('src');
-        }
-    }, {
-        trigger: '.oe_optional_products_modal .js_product:eq(1) input[data-value_name="Aluminium"]',
-    }, {
-        trigger: '.oe_optional_products_modal .js_product:eq(1) div:contains("Conference Chair (Aluminium)")',
-        run: function () {
-            var newVariantImage = $('.oe_optional_products_modal .js_product:eq(1) img.variant_image').attr('src');
-            if (newVariantImage !== optionVariantImage) {
-                $('<p>').text('image variant option src changed').insertAfter('.oe_optional_products_modal .js_product:eq(1) .product-name');
-            }
-        }
-    }, {
-        extra_trigger: '.oe_optional_products_modal .js_product:eq(1) div:contains("image variant option src changed")',
-        trigger: '.oe_optional_products_modal .js_product:eq(1) input[data-value_name="Steel"]',
-    }, {
-        trigger: 'li.js_attribute_value span:contains(Aluminium)',
-        extra_trigger: '.oe_optional_products_modal',
-        run: 'click'
-    }, {
-        trigger: '.oe_price span:contains(22.90)',
-        run: function (){}, // check
-    }, {
-        trigger: '.oe_optional_products_modal .js_product:has(strong:contains(Conference Chair)) .js_add',
-        extra_trigger: '.oe_optional_products_modal .js_product:has(strong:contains(Conference Chair))',
-        run: 'click'
-    }, {
-        trigger: '.oe_optional_products_modal .js_product:has(strong:contains(Chair floor protection)) .js_add',
-        extra_trigger: '.oe_optional_products_modal .js_product:has(strong:contains(Chair floor protection))',
-        run: 'click'
-    }, {
-        trigger: 'span:contains(1,269.80)',
-        run: function (){}, // check
-    }, {
-        trigger: 'button:has(span:contains(Proceed to Checkout))',
         run: 'click',
     }, {
         trigger: 'span:contains(Custom: Wood)',
