+++ addons/website_sale_product_configurator/views/assets.xml
@@ -11,6 +11,8 @@
         </xpath>
         <xpath expr="script[last()]" position="after">
             <script type="text/javascript" src="/website_sale_product_configurator/static/src/js/website_sale_options.js"></script>
+            <script type="text/javascript" src="/website_sale_product_configurator/static/src/js/website_sale_tour_buy.js"></script>
+            <script type="text/javascript" src="/website_sale_product_configurator/static/src/js/website_sale_tour_shop_custom_attributes_value.js"></script>
         </xpath>
     </template>
 </odoo>
